tests/Document-getElementById.html

<!DOCTYPE html>
<title>Document.getElementById with an empty string</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="../web-dom-core#dom-document-getelementbyid">
<link rel="help" href="../web-dom-core#concept-id">
<script src="testharness.js"></script>
<body>
<div id="log"></div>
<div id=""></div>
<script>
test(function() {
  assert_equals(document.getElementById(""), null)
  // XXX needs more tests
})
</script>
